+++ packages/hello/build.sh
@@ -1,4 +1,4 @@
 #!/bin/sh
 set -e
 mkdir -p "$out/bin"
-$CC hello.c -I "$LIBGREET/include" -L "$LIBGREET/lib" -lgreet -o "$out/bin/hello"
+$CC hello.c -I "$LIBGREET/include" -L "$LIBGREET/lib" -Wl,-rpath,"$LIBGREET/lib" -lgreet -o "$out/bin/hello"